Enable MongoDB and refactor TransactionPersistence implementation
EnableMongoRepositories annotation was added to AccountInquireServiceApplication to work with MongoDB. The TransactionRepositoryImpl was renamed to TransactionPersistenceImpl to better reflect its purpose and its dependency on TransactionRepository was changed from constructor injection to setter injection. The @Repository annotation was removed from the TransactionRepository interface as it is unnecessary when using @EnableMongoRepositories. The TransactionEntity class was also updated to include @Document annotation to mark it as a MongoDB document. Furthermore, the TransactionMapper utility class was updated to prevent instantiation. Lastly, refactor RestService to improve readability.