Add RestServiceTest and modify ClientSpecialType enum
A test class called `RestServiceTest` has been added in order to test the `RestService` under different scenarios, enhancing code reliability. Also, the `ClientSpecialType` enum was updated with an additional enum value `NONE` for cases where there is no special classification for a client and to ensure that the order of the enum elements is according to Java enum conventions. This would improve flexibility and readability.